src/backend/django_backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyascii==0.3.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.7
django-cors-headers==4.6.0
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
docopt==0.6.2
dotmap==1.3.30
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
frozenlist==1.4.1
gunicorn==23.0.0
idna==3.7
Jinja2==3.1.4
l18n==2021.3
laces==0.1.1
MarkupSafe==2.1.5
mjml==0.11.0
multidict==6.0.5
mysqlclient==2.2.6
openpyxl==3.1.5
packaging==24.2
pillow==10.4.0
pillow_heif==0.18.0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
telepath==0.3.1
typing_extensions==4.12.2
urllib3==2.2.2
wagtail==6.1.3
Willow==1.8.0
yarl==1.9.4
